Track the neutral mtp:: API in the CLI
Update `mtp-rs-cli` for the backend-neutral library API:
- Rewrite `cli/error.rs` to map the neutral `mtp::Error` variants to CLI exit
  codes/messages (the old `Error::Protocol { code }` / `Error::Usb` / `Error::Io`
  arms are gone). macOS USB user-client denial is now detected via the message
  text the neutral error preserves.
- `download_stream(h)` → `download(h, ByteRange::Full)` in `get` and `put --verify`.
- Source `StorageRow`/`ObjectRow` from neutral fields (`total_capacity`,
  `free_space`, `is_writable`, `ObjectFormat::code()`) while keeping the JSON
  field names stable; `handle`/`storage_id` columns become `u64`.
- `reset.rs` keeps using low-level `PtpDevice`/`NusbTransport`, now converting
  their `PtpError` into the neutral `Error` for the shared error formatter.

+++ crates/mtp-rs-cli/src/cli/commands/mkdir.rs
@@ -12,7 +12,7 @@ struct MkdirRow {
     operation: &'static str,
     remote_path: String,
     filename: String,
-    handle: u32,
+    handle: u64,
 }
 
 pub async fn run(cli: &Cli, args: &RemotePathArg) -> Result<(), CliError> {
